Fixed alignment issue in generation. Caused by adding special tokens at the bos by the tokenizer
